+++ simapp/v2/simdv2/cmd/root_di.go
@@ -13,21 +13,42 @@ import (
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
 	"cosmossdk.io/runtime/v2"
+	serverv2 "cosmossdk.io/server/v2"
+	"cosmossdk.io/server/v2/cometbft"
 	"cosmossdk.io/simapp/v2"
+	basedepinject "cosmossdk.io/x/accounts/defaults/base/depinject"
+	lockupdepinject "cosmossdk.io/x/accounts/defaults/lockup/depinject"
+	multisigdepinject "cosmossdk.io/x/accounts/defaults/multisig/depinject"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/config"
 	nodeservice "github.com/cosmos/cosmos-sdk/client/grpc/node"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	"github.com/cosmos/cosmos-sdk/std"
 	"github.com/cosmos/cosmos-sdk/x/auth/tx"
 	authtxconfig "github.com/cosmos/cosmos-sdk/x/auth/tx/config"
 	"github.com/cosmos/cosmos-sdk/x/auth/types"
 )
 
-// NewRootCmd creates a new root command for simd. It is called once in the main function.
-func NewRootCmd[T transaction.Tx]() *cobra.Command {
+// NewCometBFTRootCmd creates a new root command for simd,
+// using the CometBFT server component for consensus.
+// It is called once in the main function.
+func NewCometBFTRootCmd[T transaction.Tx]() *cobra.Command {
+	return NewRootCmdWithConsensusComponent(func(cc client.Context) serverv2.ServerComponent[T] {
+		return cometbft.New[T](
+			&genericTxDecoder[T]{cc.TxConfig},
+			initCometOptions[T](),
+			initCometConfig(),
+		)
+	})
+}
+
+// NewRootCmdWithConsensusComponent returns a new root command,
+// using the provided callback to instantiate the server component for the consensus layer.
+// Callers who want to use CometBFT should call [NewCometBFTRootCmd] directly.
+func NewRootCmdWithConsensusComponent[T transaction.Tx](
+	makeConsensusComponent func(cc client.Context) serverv2.ServerComponent[T],
+) *cobra.Command {
 	var (
 		autoCliOpts   autocli.AppOptions
 		moduleManager *runtime.MM[T]
@@ -37,19 +58,16 @@ func NewRootCmd[T transaction.Tx]() *cobra.Command {
 	if err := depinject.Inject(
 		depinject.Configs(
 			simapp.AppConfig(),
-			runtime.DefaultServiceBindings(),
-			depinject.Supply(log.NewNopLogger()),
 			depinject.Provide(
-				codec.ProvideInterfaceRegistry,
-				codec.ProvideAddressCodec,
-				codec.ProvideProtoCodec,
-				codec.ProvideLegacyAmino,
 				ProvideClientContext,
-			),
-			depinject.Invoke(
-				std.RegisterInterfaces,
-				std.RegisterLegacyAminoCodec,
-			),
+				// inject desired account types:
+				multisigdepinject.ProvideAccount,
+				basedepinject.ProvideAccount,
+				lockupdepinject.ProvideAllLockupAccounts,
+
+				// provide base account options
+				basedepinject.ProvideSecp256K1PubKey),
+			depinject.Supply(log.NewNopLogger()),
 		),
 		&autoCliOpts,
 		&moduleManager,
@@ -83,12 +101,12 @@ func NewRootCmd[T transaction.Tx]() *cobra.Command {
 		},
 	}
 
-	initRootCmd(rootCmd, clientCtx.TxConfig, moduleManager)
+	consensusComponent := makeConsensusComponent(clientCtx)
+	initRootCmd(rootCmd, moduleManager, consensusComponent)
 
 	nodeCmds := nodeservice.NewNodeCommands()
 	autoCliOpts.ModuleOptions = make(map[string]*autocliv1.ModuleOptions)
 	autoCliOpts.ModuleOptions[nodeCmds.Name()] = nodeCmds.AutoCLIOptions()
-
 	if err := autoCliOpts.EnhanceRootCommand(rootCmd); err != nil {
 		panic(err)
 	}