+++ simapp/v2/simdv2/main.go
@@ -12,7 +12,7 @@ import (
 )
 
 func main() {
-	rootCmd := cmd.NewRootCmd[transaction.Tx]()
+	rootCmd := cmd.NewCometBFTRootCmd[transaction.Tx]()
 	if err := serverv2.Execute(rootCmd, clientv2helpers.EnvPrefix, simapp.DefaultNodeHome); err != nil {
 		fmt.Fprintln(rootCmd.OutOrStderr(), err)
 		os.Exit(1)